Update Polynomial and Polyroots handling
Polynomial.hpp was updated to handle the case when the divisor of a polynomial does not have coefficients or when the order of the divisor is larger than the dividend.

The testPolynomials.cpp was modified removing the commented out test case "nxx::poly - Polynomials with real coefficients and real roots", restructuring test cases and adding additional test cases related to the polynomial root calculations.

Polyroots.hpp was also modified to order roots based on tolerance. Minor bugs related to root determination were fixed. It also included TODOs for future improvement points in the polishing, efficiency, and handling of the roots.