Remove unnecessary PolynomialError class and improve safety
Removed the PolynomialError class as it was unnecessary and was complicating the code unnecessarily. Added compile definition NOMINMAX in the CMakeLists.txt file. Enhancing safety by adding assertions in various parts of the code to prevent out of bounds. Changed the Polynomial division and root finding functions to throw specific NumerixxError instead of generic runtime_error, providing clearer error messages. Pulled out calculation of the square root of the tolerance in the processRoots function to not repeat that operation. Modified the processRoots function to use std::transform instead of std::for_each making it more succinct and understandable.

Reasons for deletion is code simplification and enhancing readability. The reasons for changes are to improve safety, performance, and error reporting.