"Improved robustness and accuracy of polynomial root finding methods
The commit increases the precision of the error tolerance constant EPS from 1E-5 to 1E-6 in testPolynomials.cpp, which allows the testing conditions to be more rigorous. The Catch2 framework in the testing suite was replaced by FetchContent for better compatibility and flexibility, which was reflected in the updates in vcpkg.json and CMakeLists.txt. Significant changes were made in the implementation of Laguerre's method in Polyroots.hpp, including better handling of corner cases, employing structured binding for readability, and stabilizing choice of initial guess for better convergence. The main driving force behind these changes was to improve the stability and reliability of the numerical methods used in polynomial root finding, especially in edge cases."