Refactor error handling in derivatives computation
In the Derivatives.hpp file, error handling was refactored to give more informative error messages when computation of derivatives fails. More specifically, DerivativeError was replaced with a template structure DerivErrorData storing additional data such as values and respective derivatives, which will provide more context to the error.

nlohmann_json dependency was removed from the CMakeLists.txt as it's no longer required.

In the Polyroots.hpp, a comment was added suggesting a refactoring to use the roots::fdfsolve function. This was not implemented in this commit but is left for future consideration.