Refactor numerical methods into cascade of Lambdas
Remodeled the numerical methods used for derivative and integral calculations into a cascade of extensible lambdas. This change simplifies the code structure, minimizes redundancy, and adds flexibility for extension and further modification. The direct application of lambda functions simplifies the handling of function call operators while still enabling the precision control of step sizes within the procedures. This refactoring also fits better in the general structure of the numerixx library.