Update derivative calculations in numerixx library
This commit makes several changes focused on enhancing derivative calculations. It includes an adjustment to the step size in Constants.hpp, the addition of a new pair product function in DemoDeriv.cpp, and substantial changes to Derivatives.hpp to refine calculation techniques. The latter includes the implementation of forward and backward Richardson extrapolation, as well as various point methods to improve the accuracy of derivatives.